pyproject.toml
[project]
name = "aoi-generator"
version = "0.1.0"
description = ""
authors = [
    {name = "ayo",email = "sa@risigner.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib (>=3.10.3,<4.0.0)",
    "scikit-image (>=0.25.2,<0.26.0)",
    "scikit-learn (>=1.7.1,<2.0.0)",
    "opencv-python (>=4.12.0.88,<5.0.0.0)",
    "pillow (>=11.3.0,<12.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
